guard index drops in release migration

+++ misc/phinx/migrations/20250315000000_release_artist_tag.php
@@ -22,7 +22,9 @@ class ReleaseArtistTag extends AbstractMigration {
         if ($this->hasTable('torrents_artists')) {
             $this->execute("RENAME TABLE torrents_artists TO release_artist");
             $this->execute("ALTER TABLE release_artist DROP PRIMARY KEY");
-            $this->execute("ALTER TABLE release_artist DROP INDEX GroupID");
+            if ($this->fetchRow("SHOW INDEX FROM release_artist WHERE Key_name = 'GroupID'")) {
+                $this->execute("ALTER TABLE release_artist DROP INDEX GroupID");
+            }
             $this->execute("ALTER TABLE release_artist CHANGE GroupID release_id int(10) NOT NULL");
             $this->execute("ALTER TABLE release_artist ADD INDEX release_id (release_id)");
             $this->execute("ALTER TABLE release_artist ADD PRIMARY KEY (release_id, ArtistID, Importance)");
@@ -35,7 +37,9 @@ class ReleaseArtistTag extends AbstractMigration {
         if ($this->hasTable('torrents_tags')) {
             $this->execute("RENAME TABLE torrents_tags TO release_tag");
             $this->execute("ALTER TABLE release_tag DROP PRIMARY KEY");
-            $this->execute("ALTER TABLE release_tag DROP INDEX GroupID");
+            if ($this->fetchRow("SHOW INDEX FROM release_tag WHERE Key_name = 'GroupID'")) {
+                $this->execute("ALTER TABLE release_tag DROP INDEX GroupID");
+            }
             $this->execute("ALTER TABLE release_tag CHANGE GroupID release_id int(10) NOT NULL DEFAULT '0'");
             $this->execute("ALTER TABLE release_tag ADD INDEX release_id (release_id)");
             $this->execute("ALTER TABLE release_tag ADD PRIMARY KEY (TagID, release_id)");
@@ -46,7 +50,9 @@ class ReleaseArtistTag extends AbstractMigration {
 
         // update votes table
         if ($this->table('torrents_tags_votes')->hasColumn('GroupID')) {
-            $this->execute("ALTER TABLE torrents_tags_votes DROP INDEX GroupID");
+            if ($this->fetchRow("SHOW INDEX FROM torrents_tags_votes WHERE Key_name = 'GroupID'")) {
+                $this->execute("ALTER TABLE torrents_tags_votes DROP INDEX GroupID");
+            }
             $this->execute("ALTER TABLE torrents_tags_votes CHANGE GroupID release_id int(10) NOT NULL");
             $this->execute("ALTER TABLE torrents_tags_votes ADD INDEX release_id (release_id)");
             if ($canFk) {
@@ -72,7 +78,9 @@ class ReleaseArtistTag extends AbstractMigration {
         }
 
         // revert votes table
-        $this->execute("ALTER TABLE torrents_tags_votes DROP INDEX release_id");
+        if ($this->fetchRow("SHOW INDEX FROM torrents_tags_votes WHERE Key_name = 'release_id'")) {
+            $this->execute("ALTER TABLE torrents_tags_votes DROP INDEX release_id");
+        }
         $this->execute("ALTER TABLE torrents_tags_votes CHANGE release_id GroupID int(10) NOT NULL");
         $this->execute("ALTER TABLE torrents_tags_votes ADD INDEX GroupID (GroupID)");
         if ($canFk) {
@@ -81,7 +89,9 @@ class ReleaseArtistTag extends AbstractMigration {
 
         // release_tag -> torrents_tags
         $this->execute("ALTER TABLE release_tag DROP PRIMARY KEY");
-        $this->execute("ALTER TABLE release_tag DROP INDEX release_id");
+        if ($this->fetchRow("SHOW INDEX FROM release_tag WHERE Key_name = 'release_id'")) {
+            $this->execute("ALTER TABLE release_tag DROP INDEX release_id");
+        }
         $this->execute("ALTER TABLE release_tag CHANGE release_id GroupID int(10) NOT NULL DEFAULT '0'");
         $this->execute("ALTER TABLE release_tag ADD INDEX GroupID (GroupID)");
         $this->execute("ALTER TABLE release_tag ADD PRIMARY KEY (TagID, GroupID)");
@@ -92,7 +102,9 @@ class ReleaseArtistTag extends AbstractMigration {
 
         // release_artist -> torrents_artists
         $this->execute("ALTER TABLE release_artist DROP PRIMARY KEY");
-        $this->execute("ALTER TABLE release_artist DROP INDEX release_id");
+        if ($this->fetchRow("SHOW INDEX FROM release_artist WHERE Key_name = 'release_id'")) {
+            $this->execute("ALTER TABLE release_artist DROP INDEX release_id");
+        }
         $this->execute("ALTER TABLE release_artist CHANGE release_id GroupID int(10) NOT NULL");
         $this->execute("ALTER TABLE release_artist ADD INDEX GroupID (GroupID)");
         $this->execute("ALTER TABLE release_artist ADD PRIMARY KEY (GroupID, ArtistID, Importance)");
